[Bug 781721] Add browser detection function.
Detect platform and Firefox version based on a useragent string. This is
provided as a function and as a middleware that can automatically put
the values in request.BROWSER.

This makes no attempt to try and work with non-Firefox browsers. It
does not support Fennec, but does properly detect the new Firefox
for android with version >14.