[Bug 781718] Dashboard structure, recent feedback.
Add some structure to the dashboard page and a list of recent feedbacks.
This still needs to bring in a lot of the other features of the
dashboard, such as sorting and filtering, but it is a useful first step.

Unlike other front end code, this was not migrated from old input, but
was written from scratch to look very similar.