Add sampledata command
This gives us a structure for easily adding sample data in an instance
which will make it easier for developers to get up to speed quickly.

This uses the same code that Jingo uses for importing a module that
might not exist, but in such a way that it doesn't mask valid ImportError
problems.