Implement Dashboard
The dashboard has a histogram that shows amount of input over time and a
filter/facet sidebar to drill down on data. The selected feedback items
are displayed below the graph. This is a fairly faithful reproduction of
old input, but written from scratch.

This uses Flot to draw the graphs, and ES to generate data. There are no
SQL queries made in generating this page.