[bug 791381] Fix dashboard regarding ES errors
This fixes the dashboard page so that if there are any
ElasticSearch-related errors, it shows the es_down.html template.

In doing this, I tweaked the structure of the dashboard.html template
so that I could extend it for the es_down.html template. I also
pulled in a new version of ElasticUtils.