Replace fjord.sampledata with django-eadred
I spent some time extracting the sampledata application out as a separate
project so we could use it in Kitsune (and other places as well). This
switches fjord over to using that.

One nice thing eadred has is the ability to pass parameters from the
command line to the generate_sampledata functions. This allows us to
easily allow for different kinds of sample data generation: reproducible
data set for contributors, randomized large data set for testing the
dashboard, icky data set for testing security and unicode handling, ...