[Bug 722388] Collect emails from users.
This adds a checkbox and a input box to the survey form. The checkbox is
to indicate the user wants to give an email address, and the input field
is for the email address, and only shows up after the box has been
checked.

This email address is stored in a seperate table with a foreign key, for
possible security benefits down the road.