[Bug 799693] Zero fill histogram data.
Fill the graph with data at every day along the X axis, instead of
allowing missing data to be interpolated. Care is taken to make sure
that these extra zero points align to the spacing of the rest of the
points (24 hours per point).

Other minor, related changes:

- If start or end is not specified, they have default values now.
- The infinity date selector was replaced with a 90 day button.
- The one day selector now hides the graph.
- Due to the fixed, relative time frame, tests have been changed to use
  relative dates.

The zero fill algorithm benefits from having non-None start and end
times, and these changes help with that.