[Bug 799691] The histogram data is in UTC.
This makes the graph lines and the graph data points line up. This only
changes flot's expectations to match what elastic is ending it.

When elasticsearch returns histogram data with a period of 1 day (like
we are doing on the histogram) the timestamps it returns have the hours
of a UTC midnight date, regardless of what timezone the aggregate set
represents.

This means that when the server is working in Pacific time, it
aggregates a Pacific day worth of data, sums it up, and then returns it
labeled with a timestamp. This timestamp, when output, is the midnight
of a UTC day (calculating the hours past midnight only returns 0 if you
assume UTC), even if localtime on the server is Pacific.

Even if we change to forcing a particular timezone in the future, this
will still be true, and elasticwill still label this data with UTC,
regardless of the input.

Because of this, flot needs to be told the data is UTC, not browser
local time.