[Bug 805020] Don't checkout any branch on vendor.
Vendor is a gitsubmodule, which means we should be obeying git's idea of
what commit it should be on. We should not checkout out a particular
branch, because this can cause changes to come in to early, such as the
celery 2.6 upgrade, which caused bug 805020.

Additionally, I added --recursive to the submodule checkout in the main
repo, since we aren't doing this explicitly in vendor anymore.