[Bug 832037] Update for pyelasticsearch-based elasticutils
This updates the codebase to use the pyelasticsearch-based elasticutils
from master tip.

Generally the changes are pretty minor. Most of them are to handle
the pyelasticsearch exceptions rather than the pyes ones. A few are
because some of the API bits changed in elasticutils. Then there's
some very minor cleanup.