[bug 855671] Handle phony addon using FfA feedback
Firefox for Android users could be using the phony addon which changes
their user agent to something else. The problem being that our code
would throw an exception because various fields in the feedback
response table can't have nulls.

This changes how we parse user agent strings so that instead of None
we now use "Unknown".

This lets Firefox for Android phony users to enter in feedback and
we'll capture the phony user agent, keep the feedback and attribute
it to "Unknown".