Fix version filter
If the user isn't filtering on product, then we should ignore the version
filter.

The better fix would also nix it from the url, but that's a lot harder
and requires some non-trivial rewiring so I'm going to do this for now
and if that's not sufficient then figure out a different fix.