[bug 793850] Add product to mapping
This adds a new field called "product" to the mapping. This holds the
name of the product which is based on the browser and platform. This
assumes that "Firefox for Android" is always "Firefox" browser on
the "Android" platform.

This fixes tests accordingly.

This adds Firefox for Android to the sampledata generator.