[bug 878206] Fix happy hardcoding horseshit
This un-hardcodes Firefox for Android responses and sets the happy
value based on the _type value correctly.

Note: Our tests were wrong before: 1 is happy, 2 is sad and 3 is idea,
but our tests were coded such that 1 was happy and 0 was sad or
something like that. This fixes that, too.