Nix the DEV_LANGUAGES thing
It's annoying as all hell to have DEV_LANGUAGES and PROD_LANGUAGES where
one is generated and the other is set in stone and you see different
things depending on whether you're in a dev environment or not. It
essentially means you could be donig some work, all tests pass, you
push it to master and then it dies on stage or prod because those
are different.

This nixes that and reduces the amount of misery in our lives.
Now there is only PROD_LANGUAGES.