Add second data migration to fix responses
The 0008 data migration takes a long time to run. It missed a bunch
of responses which I think were created after the sql query was
run, so the ids weren't included in the batch.

We could fix the 0008 data migration, go back and run it again, but
it's way easier to just write a new data migration to patch the
situation.