[bug 905372] Change 'Unknown' to ''
This reworks the data production side of Fjord so that it uses ""
instead of "Unknown" when it can't infer things. This puts it in line
with the API defaults which are "" and thus should make the data more
consistent between data produced by the API and data produced by the web
form.

At this time, there are only 395 responses that have "Unknown" in them.
The migration will fix those to "" and we'll use "" going forward.

The wry twist is that "" is a lousy value for the ui, so for ui display
purposes, we convert "" to "Unknown".