[bug 915695] Infer FxOS version from Gecko version
This adds some code to infer the FxOS version from the Gecko
version specified in the browser parts of the user agent. This
isn't a perfect way to get the version--there's a bunch of data
that doesn't have the right Gecko versions. However, it's
probably accurate for certain Gecko versions.

This includes a data migration that backfills the data.