[bug 913613] Add wave 2 countries, fix ui
The previous ui for choosing countries didn't scale well for lots of
countries. So I switched it to a select box. I switched the devices
chooser to a select box, too, figuring fewer kinds of ui inputs the
better.

This also adds wave 2 countries that have shown up in public
announcements.