[bug 929664] Truncate description at 10000
This truncates the description at 10,000 characters. I'm doing it in
code-space rather than database-space because we're probably going to
tweak the value a bit as time goes on and it's easier to do that in
code.

This includes a data migration to fix all the existing responses. In
my db, this fixes 163 of them.