[bug 918959] Strip that whitespace!
This tweaks Response.save() to strip description whitespace before
saving to the db. We can do this in other places, but given that all
those other places feed in through here, it makes sense to do it here.

This also includes a data migration to strip all the existing responses.