[bug 936625] Add country to serializer
This adds the country to the serializer so it gets captured by the API
and stored in the db.

This adds a test that wouldn't have found this problem, but is a good
test to have anyhow since it makes sure all the pieces line up and that
we're testing a "maximal" API call.