[bug 902032] ratelimit against double-submit
This ratelimits against double-submits by ratelimiting to 1/m on ip
address plus first 50 characters of the description.

Also, I moved the actual_ip key function to the feedback view. It's not
going to be used anywhere else, so it seems silly to have it far away
from where it's going to be used.

Updated the tests to work with the new double-submit avoidance code.

I also wrote a new ratelimit decorator that allows us to statsd log
helpful bits in a stack-friendly way and also uses Fjord-friendly
defaults.