[bug 927617] Add occurrence comparison report
This adds a prototype of the occurrence comparison report. It uses
ElasticSearch to prep text into a set of tokens, generates bigrams
and then uses ElasticSearch terms facet to compare bigrams between
two sets of things.

This was supposed to do trigrams, too, but trigrams aren't working with
ElasticSearch 0.20.5. It works with 0.90.7. I'm not entirely sure what
the problem is so pushing that off to a separate task.