[bug 947278] Tweak actual_ip_plus_desc
I saw a couple of errors from production which kicked up a
"MemcachedKeyLengthError: Key length is > 250" error. I tried to create
a test using data from the errors, but wasn't able to reproduce the
problem.

While doing that, I decided to write a test for actual_ip_plus_desc to
vaguely verify correctness specifically with unicode descriptions. I
also tweaked the algorithm so it's less likely to produce a key that's >
250 characters.