[bug 947278] Switch memcached key generator to use md5
This reworks the memcached key generator to use md5 to hash the
description and then attach that to the ip address rather than the other
thing I was doing.

Hopefully this nixes the problems I'm seeing in production with keys
that exceed memcached's 250 character limit, though I'm still not
totally clear on why that's happening.