[bug 931189] Add character counter to mobile feedback form
This adds a character counter to the mobile feedback form. This doesn't
prevent users from submitting the form--it just indicates how many
characters are acceptable, when the user is getting close to the limit
and when the user is over the limit.

Also, I moved part of the privacy text upwards so it's right under the
description. This is not a change that requires l10n work. It is a
change that the Privacy folks suggested in bug 934721 for the other
feedback form.