[bug 857056] Rework feedback urls
This reworks feedback urls so that we can take the locale, product,
version and channel from the url and fallback to inferring it from the
user agent.

This allows products to use product-specific urls which gives us the
possibility of capturing the channel and allows us to more correctly
distinguish Firefox, Firefox for Android, Firefox OS and Firefox Metro
(or whatever the name for that will be).

If an unknown product is specified in the url, then the "unknown product"
page will show.

This also adds a bunch of FIXME comments for things that should get fixed
in future commits some of which are hard-coded right now. Why in the
future? Becuase they're problems that have other pieces and should be
thought through before committing to code and creating new
backwards-compatability issues.