[bug 933786] Fix strip error
DRF allows nulls as valid values for CharField. This creates
NoNullsCharField which restores sanity to the world.

This fixes the strip error by raising a ValidationError which will send
an error to the client rather than letting the value through and having
it throw an AttributeError where it shouldn't be.