Rework histogram-less situation
The es_down template kicks in if ES is having problems. It derives from
the dashboard template which includes JS for setting up the controls on
the dashboard.

This reworks some of that code so that when es_down is showing, it won't
include the dashboard js anymore and thus we can nix some of that "if no
histogram, then no show" code.