[bug 982242] Add description_terms field
This adds a new field to the index that has a different analyzer than
the existing description field. This one doesn't do stemming so it's
good for the terms facet.

Wait, what? Yeah, so when Elasticsearch does stemming, it stores things
in the index like "becaus" instead of "because". That's great for
searching, but lousy for doing other things with the resulting tokens
like word clouds, comparing sets of bigrams, etc.