Fix messages text
There's a non-trivial number of people who are completely confused about
the front page dashboard graph and the messages number to the right of
it. It seems they think that the messages number is a total of all
messages in the system and that we're removing responses thus the number
is going down.

That's not true at all. What's going on is that it's the number of responses
in the datetime range specified. As time goes on, that range moves and thus
the number goes up and down depending on how many responses there are in
the range. We have like 680k responses in the db right now and that
number is going up monotonically.

This string tweak hopefully alleviates the confusion for some people.