[bug 1003262] Add api column
Previously we were (ab)using the user_agent column to denote feedback
responses that came in via the api. This fixes that. Now we have an api
column that specifies the version of the api that was used to generate
the feedback response.

This also adds a data migration fixing the existing responses that had
"api" in the user_agent field. Now they'll have 1 in the api field.

This also adds a facet filter to the analyzer search view partially so
that I can more easily keep track of how often the api is getting used
and by what.