[bug 949101] Add mlt section to response view
This adds a morelikethis section to the response view. It's "prototype"
quality. I had a hard time fine-tuning the MLT query and ended up just
adding a massive stopwords list which seemed to reduce the number of
"wow--that has nothing to do with this" responses. This might benefit
from synonyms if we ever add that.

This also adds a test for the response view to make sure the "secret"
section only shows up for analyzers.