[bug 1007826] Remove dashboard link
The dashboard link leads users to a meh dashboard that's confusing at
best and easily misinterpreted. Amongst other things, the link reads
"Want to see where your feedback goes?" which suggests the dashboard is
a destination for feedback when it's just a (lousy) view of the
feedback. It doesn't talk about any of the systems we use to aggregate
and examine the feedback.

This removes the link for now. When the new dashboard system lands, we
can reinstate it.

This also removes some dead code which never executes since "opinion"
is never defined.