[bug 987693] Switch to grappelli
This is a more aesthetically pleasing admin theme that more closely
matches the ambiance in my office so we're switching to it.

In order to get django-adminplus to work with grappelli, we had to add a
tiny bit to the admin/index.html template. This was easiest to throw in
a separate app so that we could make sure that app shows up before
grappelli in the INSTALLED_APPS list. This app is appropriately known as
grappellioverride. It is being optioned for a major motion picture soon.