[bug 971369] Automated Gengo machine translation
* adds automated Gengo machine translator
* adds Gengo API wrapper
* adds tests
* adds some statsd calls so we can sort of see what's going on

The tests are a little goofy since they're very mocked so as to not
require public and private keys to run.

This infrastructure doesn't work at all unless you have working public
and private keys from Gengo. If you don't, then when it goes to auto
translate, it'll throw a GengoConfigurationError.

This is a first pass. If we can't guess the language of the text, then
we don't do anything with it. Further, if there's an error with machine
translation, we don't do anything more with it.