Remove vendor-fixing from deploy script
A while back, we switched from vendor/ as a git submodule to vendor/ as
a directory with a bunch of git submodules. To do that "cleanly" we had
to do a "destroy all the things with fire" rm -rf thing in the deploy
script. We don't need that anymore.

Further, removing that, should speed up deployments.