[bug 933390] Add input-tests code
This adds the input-tests code from the willkg/input-tests repository
based on sha 5d2a05b.

Adding it to the fjord repository lets us tie versions of fjord to
changes/updates in the smoketests. This isn't designed to replace the
input-tests repository so much as it's designed to gives us
client-based smoketests which we're sorely lacking right now.