Fix problem where existing description scenario fails
If the user tried to submit a response and wasn't able to, we store all
the data in storage for the user to try again later. If the user tries
again later, this will kick up an error instead of going to the last
card so the user has to flip through all the cards again in order to try
to resubmit.

This fixes that.