[bug 1033852] Better handle unsupported src languages
Adds some code to not try known unsupported languages with the Gengo
machine translator. The human translator has a way to find out what
language pairs are supported--the machine translator does not. Thus we
hard-code the list of known unsupported languages so we don't keep
trying and kicking up errors.