[bug 1016466] Remove funfactory
Funfactory... It's not you, it's me. We're just not good for each other
anymore. I look to the future, and it's a future of double rainbows and
misty sunrises. It's a future where I've got more control and in turn
more responsibility. It's a future where I'm empowered to have my own
successes, and my own mistakes. Maybe I'll double the Firefox userbase.
Maybe I'll become the cornerstone of a bad news cycle (again). I don't
know for certain what will happen, but I know it needs to happen without
you.

Be well, Funfactory. I hope we can still be friends.