[bug 987801] Exit if requirements are mismatched
This adds some code that gets run via ./manage.py which will cause it to
exit if the dependencies specified in the requirements file don't match
what's installed in the virtual environment.

It can't effectively check requirements specified by github tarball url.
For those, it puts them in another list and then tells you how many it
can't check. Going forward, we should reduce the number of these because
we can't do a lot about them.

It lets you quell the check with SKIP_CHECK=1.