[bug 987801] Redo how we figure out whether to use vendor/
This adds a line so that it automatically activates the virtual
environment when you do "vagrant ssh". This makes "./manage.py" work
fine as well as the code for determining whether requirements are
mismatched.

This changes fjord/manage.py such that it (ab)uses the fact we're in
vagrant and know wehre the virutal environment is to figure out whether
to enable the vendor/ stuff or not. This is just a stopgap until we
remove vendor/ altogether.