Smoketest cleanup
Since it's now unlikely that QA will ever use this code, I reorganized
the files so it's closer to how Input is structured:

* got rid of desktop/ intermediary directory--we don't really
  differentiate between desktop and mobile
* changed FeedbackPage to DashboardPage which is a more proper name for
  it
* nixed BasePage--that was supposedly the dashboard page, but it was
  goofypants so it needed to be shown the door
* tweaked some code to function better and faster

Generally, this is cosmetic in the sense that it doesn't affect what
we're testing and how we're testing it--it's just code organization and
some speed improvements. Went from 160s to about 120s runtime.