[bug 1055785] Codify data retention policy
This adds the data retention policy to the About page.

It also nixes the mobile template for the About page which wasn't
very functional.

It also adds text and links which are already in the footer, but this
surfaces them so that the About page is "complete" in terms of the
site's policies.