[bug 1055788] Implement data purging
This implements the data retention policy from bug #946456:

* delete feedback_responseemail records older than 6 months
* delete feedback_responsecontext records older than 6 months

It tries to do the work in bulk so as to minimize the db traffic but
I'm pretty sure the .delete() is doing dumb things like getting a list
of ids then deleting the rows in chunks. I look forward to updating to
Django 1.6 and seeing if it still happens like that.