[bug 1062453] Add flagged view
This adds a really goofy looking flagged view to the analyzer's
dashboard. It shows en-US responses and their flags and allows you to
toggle the "spam-wrong" and "abuse-wrong" flag. It also filters on tags
so we can sort of see what's going on.

It's probably the case this view needs some additional things, but we
can add those later.

Also, I did a lazy data migration to add the initial set of flags.