Update peep and codify --no-use-wheel
After working on bug #1074861 (peep issue #50), I decided we shouldn't use
wheels for now. The crux of the issue is that if you have a github-tarball
based requirement and then switch to a pypi based requirement which
downloads a wheel, then pip doesn't clean things up properly and you end
up with cruft in your virtual environment which includes .py, .pyc and
metadata files. Ew.

Further Mike and Dean said they've had instances where peep downloads a
wheel and then a non-wheel, so you need to have both hashes in the
requirements file. Ew.

Because of these two issues, it's easier to not do wheels. So I updated
peep to the latest and then modified it so it's always passing
--no-use-wheel to pip.

I also fixed the shas in the requirements file that were for wheels.

Plus when testing all of this, I had a problem where the vagrant
provisioning script was creating /root/.pip and dying with a permission
issue. So I fixed that, too.

* https://bugzilla.mozilla.org/show_bug.cgi?id=1074861
* https://github.com/erikrose/peep/issues/50