Redo the peep --no-use-wheel fix
pip 1.4 through 1.5.6 don't upgrade wheels correctly. I fixed our peep
version so that it handles this by not using wheels, but that fix
doesn't work with versions of pip prior to 1.5.

This fixes that by adding some pip version checking. It also centralizes
the fix into one place and documents it better.