[bug 1020303] Add rating column
This adds a rating column which is like the happy column, but allows for
a gradation of values.

If we get a response with a rating, but no happy, then we'll populate
the happy accordingly--we always want to have a happy.

However, if we get a happy, but no rating, we won't populate the rating
column--we'll leave it null.

This allows us to A/B test the two and distinguish between which version
of the form the user saw and correlate that to the answer they gave.