[bug 854479] Fix ui-lightness 404 problems
This was a tricky problem with paths between our server environments and
development environments. I moved some files around so the relative urls
that jquery-ui.css use are correct for both the minified dashboard.css
file which we use for server environments and the jquery-ui.css file we
use for development environments.