[bug 1084387] Add on_picker field to Product
If a product should show up in the product picker, then on_picker=True.
If a product should not show up in the product picker, then
on_picker=False.

In this way, we don't accidentally reveal some products we need to
gather feedback for, but don't want to confuse users leaving feedback
for other products with their existence.