Fix throttle test
This fixes the throttle test so that it clears the cache before running
the test and also pauses between HTTP requests. This reduces the
likelihood that it bumps into submitted responses from other tests. The
slight pause makes it pass consistently on my local machine.

Added a FIXME to look into the "pause" thing further because that's
total goofipants.