[bug 1097352] Response view is viewable by all
Previosuly, we were using the Product on_dashboard flag to cover the
response view as well. Bug 1097352 covers adjusting that thinking so
that the on_dashboard flag literally just dictates whether it shows up
on the front page dashboard and doesn't affect permalinks. It will add
finer-grained public/private flags.

This fixes the response view to allow all responses be public. We'll do
the rest of the work for bug 1097352 later.