[bug 1097847] Fix custom views in admin
django-adminplus changed the register_view signature. This updates all
the places we use admin.site.register_view so it's correct and our
custom views show up in the admin.

django-grappelli 2.5.6 has a section for django-adminplus, but it's
bizarre and doesn't work right. This updates our grappellioverride to
make it less disfunctional.

Updated the comment in fjord/settings/base.py regarding the grappelli
override.