[bug 1103024] Gengo kill switch
Periodically, Gengo does system maintenance and during this period they
don't want anyone doing API calls or creating translation jobs.

This creates a waffle switch that allows us to enable/disable all
Gengo-related activities across the site. If gengosystem switch is
disabled, we don't do any translations, pushes or pulls.

We're going for simplicity here. Thus the thing this doesn't cover is
what happens to all the things that *should* have been translated during
that time. For that, we have the backfill tool which is available in the
admin in the "Translations - Gengo Maintenance" view.