[bug 1111265] Restrict remote-troubleshooting to Firefox
If there's a mismatch between the product the user is leaving feedback
about and the browser the user is using, then we really don't want the
browser data. I haven't figured out how to programmatically figure that
out in a general way, yet.

In the meantime, we're going to restrict this to Firefox product only
because that's where we get the best bang for our buck right now anyhow.