[bug 1091716] Carry changes over to generic_feedback_dev.js
In bug #1091716, we updated
fjord/feedback/static/js/generic_feedback.js. However, because we're in
the middle of the remote-troubleshooting changes, we have a parallel
generic_feedback_dev.js file which has the remote-troubleshooting
changes and is currently "live".

This commit updates the generic_feedback_dev.js file with the changes we
made to generic_feedback.js.