[bug 1076879] Fix Response indexing retry
Occasionally, we'd get this error:

Task fjord.search.tasks.index_item_task with id
10b7fc0e-6f40-4338-9b4f-f449b0906676 raised exception:
'AttributeError("type object \'ResponseMappingType\' has no attribute
\'split\'",)'

What's going on is if indexing fails, then it builds a new task to retry
indexing. When it did that, it was incorrectly building the args
argument using the mapping type instance rather than the stringified
class path representation. This fixes that.