[bug 1111265] Fix browser data re: products and browsers
Previously, we had the template code hardcoded to only collect browser
data if the product the user is leaving feedback for was "Firefox".
This prevents us from collecting browser data for "Firefox Dev Edition".

This changes the code so that the Product specifies whether to collect
browser data at all and which browser we'll collect browser data from.
For example, if the product has browser_data_browser=u'Android' and the
user is using Firefox for Android, then we'll add the browser data
collection bits to the template. Otherwise, we won't.

Also, I moved the "should we show the browser data section" bits from
the template to the view and added some better documentation.