[bug 1093341] Fix gengo warning emails
We want an email warning that the Gengo account balance is low to be
sent out once per day rather than every time the sync stuff runs. This
adds a "daily activities" kind of thing to translation modules so that
we can run:

    ./manage.py translation_daily

and have that run all the things that should happen only once a day.

Additionally, we add a cron job so that this runs in the morning.