[bug 1129085] Infer version for Firefox Dev
Previously, we'd only infer the version if the browser from the user
agent matched the product db_name. That doesn't work for Firefox Dev
Edition or Firefox 64.

This fixes that so that we can explicitly state which "browser" a
product is. Then it'll match up, we can appropriately infer the
product version from the user-agent specified browser version, and we
can go home at night and feel like we did a good job at work.