[bug 1136809] Fix (person, survey, flow) uniqueness issues
This tweaks the Answer model so that (person_id, survey_id, flow_id) is
a unique tuple in the data set. In order for that to work with the fact
we use POST to update, I had to disable the uniqueness check when the
serializer calls full_clean on the model. I think that'll be ok, but if
it's not, it'll throw an error and we'll know soon enough.

This involves three migrations: one to create an index on (person,
survey, flow) so that the next migration is *MUCH* faster when it
removes the duplicates. The last migration adds the uniqueness
requirement.