Fix migration for fixing AlertFlavor.allowed_tokens
Recently, I changed AlertFlavor.allowed_tokens so that you could create
a flavor without specifying the tokens that go with it (i.e. I added
blank=True).

The resulting migration for that change does a bunch of SQL, but doesn't
actually change the db.

So I'm tweaking the last exsiting migration rather than creating a new
migration that does a bunch of stuff to do nothing.