[bug 1139545] Fix "get or create" section of HB post view
This tweaks the POST view so that it tries to create the object first.
That will fail with an IntegrityError because there's a uniqueness
constraint on (person_id, survey_id, flow_id). If it fails, then
there's already a record like this and it fetches that and updates it.

In this way, we use the db to eliminate the race condition between
the "get" and "create" steps which is kicking up the IntegrityErrors
we're seeing.