Rework commit message linter
The commit message linter looked at changed code if I used -v when doing
git commit. Further, if the commit message was rejected because of
errors, there was no way to re-edit it without shenanigans.

This reworks it so that it asks you if you want to re-edit the message.
Plus it gives you better information about where the error is. Plus it
ignores everything after the # ... stuff.