[bug 1159882] Change "happy" in API to default to False
Bug #1155600 creates the problem where Firefox OS > 1.3 relies on the
old (busted) behavior where the "happy" field wasn't required and
defaulted to False. This changes the BooleanField back to that behavior
so that Input stops rejecting Firefox OS > 1.3 feedback.

One caveat of this is that all feedback will be "sad" regardless of
whether the user was actually happy or not. That makes me sad, too.