[bug 1161144] Implement suggest framework
This implements a suggest framework that allows us to have a series of
Providers which offer suggested links to list on the Thank You page
after a user has given us feedback. Initially, we'll use this to generate
links for sad feedback based on SUMO data. We can additionally use this
for other purposes such as soliciting additional information based on
trigger words.

The world will be our oyster.