[bug 1144159] Rework celeryd code and fix mysql error
We were getting a sporadic "MySQL has gone away" error in production
which we tracked down to not having DJANGO_SETTINGS_MODULE set when the
celery django fixup code executed which meant we were getting the
default celery loader which didn't clean up db connection before working
on celery tasks. Whew.

This fixes that by moving the os.environ.setdefault() line to the top of
manage.py along with a comment.

This also redoes the celeryd.py code so that it's a copy/paste of
the original source from django-celery so it's easier to maintain going
forward if they change their code.