[bug 1133771] Rework thank you page view
This reworks the Thank You page view and tests. We continue to use the
thankyou page flag, but this allows us to specify a response_id in the
querystring to make it easier to examine and debug behavior in
production.

Plus it gets rid of the thanks_sad.html template.