[bug 1077596] Temporary histogram feedback API
This is a temporary histogram feedback API. "Temporary" in this
case means we'll deprecate it as soon as we write a better one
using Elasticsearch aggregations. As soon as we do that, we'll
deprecate this one.

There's a bunch of FIXMEs in the code. We can fix them if this proves
less temporary than I'd like.