[bug 1178440] Fix TEST_RUNNER warning
We were getting a TEST_RUNNER warning which was spurious because we're
not using the Django default test runner. However, the warning kicked
off because there were some default settings in our settings.py file
which caused the checker to think we had upgraded from Django 1.5 and
thus needed the warning.

This removes some of those settings and tweaks the code so the spurious
warning doesn't pop up anymore.