[bug 1170183] Add "npm -v" to deploy script
This adds an "npm -v" to the deploy script to verify that npm is
installed, available to the deploy script and that it's a version that
we can use.

Rehan's changes for supporting django-pipeline will switch to using it
to install all the node stuff. For now, we just need to make sure that
it'll work.