Add pytest.ini to smoketests/ directory
When we switched to py.test for fjord, we added a pytest.ini to the root
directory. This created a problem since the settings in that pytest.ini
don't work for the smoketests which also use py.test.

This fixes that by creating a pytest.ini that gets used by py.test when
running smoketests so it doesn't find the pytest.ini in the root
directory which causes it to do silly things.