[bug 1191968] Restyle product picker page
This restyles the product picker page so it's more like SUMO's. It's
not exactly like SUMO's, but it's enough like SUMO's that we can use the
product icons they use. That'll make some things easier.

Further, this adds product descriptions, but those aren't shown at the
moment because we need to l10n-ize them and we're lacking that
infrastructure.