Fix indentation of user_agent_data.json file
I added a rule to .editorconfig so this file has 2-space indents and
then reindented it. Previously, it had some lines with 3-spaces and some
with 7-spaces. Easier to just normalize it and then make the changes
I need to make.

This only affects test data, so I'm just going to check it in.